--- common/greeter.go
@@ -1,7 +1,0 @@
-package common
-
-import "fmt"
-
-func Greet(name string) string {
-	return fmt.Sprintf("Hello %v \n", name)
-}
