+++ worker.Dockerfile
@@ -3,17 +3,17 @@ FROM golang:1.11-alpine as build-env
 WORKDIR /go/src/github.com/chrisstowe/forgestatus
 
 COPY common ./common
-COPY forgestatus-worker ./forgestatus-worker
+COPY worker ./worker
 
-RUN go build -o ./goApp ./forgestatus-worker
+RUN go build -o ./workerApp ./worker
 
 FROM alpine
 
 WORKDIR /opt/app
 
 # install app
-COPY --from=build-env /go/src/github.com/chrisstowe/forgestatus/goApp .
+COPY --from=build-env /go/src/github.com/chrisstowe/forgestatus/workerApp .
 
 EXPOSE 80
 
-ENTRYPOINT [ "./goApp" ]
+ENTRYPOINT [ "./workerApp" ]
